+++ src/components/StockDetails/PredictionSlider.jsx
