+++ src/components/StockTable/StockSearchBar.jsx
