+++ src/components/StockTable/StockTableCard.jsx
@@ -1,4 +1,4 @@
-import React, { useContext } from 'react';
+import React from 'react';
 import { Box, useColorModeValue, HStack, Container } from '@chakra-ui/react';
 import StockTable from './StockTable';
