+++ src/pages/Plans.jsx
@@ -1,7 +1,7 @@
 import React from "react";
-import PageTemplate from "../components/pagetemplate/PageTemplate";
-import PaymentPlanList from "../components/PaymentPlanList";
-import ReviewCardList from "../components/ReviewCardList";
+import PageTemplate from "../components/PageTemplate/PageTemplate";
+import PaymentPlanList from "../components/PaymentPlan/PaymentPlanList";
+import ReviewCardList from "../components/PaymentPlan/ReviewCardList";
 import CustomHeading from "../components/CustomHeading";
 import {Divider, useColorModeValue} from "@chakra-ui/react";
